refactor: split model portfolio reference dtos
